Extract toJson mapper
