+++ src/infrastructure/persistence/cityMapper.js
@@ -1,0 +1,15 @@
+module.exports = {
+  toJson: fields => {
+    return {
+      name: fields[1],
+      nameAscii: fields[2],
+      location: {
+        longitude: parseFloat(fields[5]),
+        latitude: parseFloat(fields[4])
+      },
+      state: isNaN(fields[10]) ? fields[10] : "",
+      country: fields[8],
+      population: parseInt(fields[14], 10)
+    };
+  }
+};
