+++ src/infrastructure/persistence/cityRepository.js
@@ -3,26 +3,21 @@ const es = require("event-stream");
 const { filter } = require("fuzzaldrin");
 const score = require("string-score");
 const geolib = require("geolib");
-const { scoreName, scoreDistance, scoring } = require("./score");
+const { scoreName, scoreDistance, scoring } = require("../score");
+const { toJson } = require("./cityMapper");
 
 module.exports = ({ dbFile = null } = {}) => {
   const MIN_POPULATION = 5000;
   const METERS_IN_KM = 1000;
   const PRECISION_HUNDRED_METERS = 100;
 
-  const toJson = fields => {
-    return {
-      name: fields[1],
-      nameAscii: fields[2],
-      location: {
-        longitude: parseFloat(fields[5]),
-        latitude: parseFloat(fields[4])
-      },
-      state: isNaN(fields[10]) ? fields[10] : "",
-      country: fields[8],
-      population: parseInt(fields[14], 10)
-    };
-  };
+  const fetchCities = () =>
+    fs
+      .createReadStream(dbFile)
+      .pipe(es.split("\n"))
+      .pipe(es.mapSync(data => data.split("\t")))
+      .pipe(es.mapSync(toJson))
+      .pipe(es.mapSync(filterMinPopulation));
 
   const filterMinPopulation = city => {
     if (city.population > MIN_POPULATION) {
@@ -46,14 +41,6 @@ module.exports = ({ dbFile = null } = {}) => {
   };
 
   const computeScore = city => ({ ...city, score: scoring(({ scoringName, scoringDistance } = city)) });
-
-  const fetchCities = () =>
-    fs
-      .createReadStream(dbFile)
-      .pipe(es.split("\n"))
-      .pipe(es.mapSync(data => data.split("\t")))
-      .pipe(es.mapSync(toJson))
-      .pipe(es.mapSync(filterMinPopulation));
 
   return {
     findByName: async name =>
